refactor(jobs): split sweep SQL constants and recovery handlers into sweep.py
jobs/router.py crossed 2047 lines with its cap raised to match. Split
along the seam the file already had: the stale-job recovery/sweep
handlers (fail_stale_jobs, sweep_stale_vrt_assets, the presigned-staging
reapers, publish_refresh_reconciliation) and their SQL constants
(_NO_LIVE_GENERATION_JOB_SQL and friends) move to sweep.py, leaving
router.py with the plain job CRUD routes (cleanup/status/retry) plus a
façade re-exporting sweep.py's public and test-facing names so every
existing import path keeps working. router.py drops under the
router-glob gate's default 1500-line cap with no ratchet entry needed.

No behavior change: every moved function is byte-identical, and
cleanup_stale_jobs stays physically in router.py (it is what tests patch
via app.platform.jobs.router.fail_stale_jobs) so the monkeypatch targets
in test_jobs_router.py and test_tenant_job_recovery.py still intercept
the right call. test_stale_pending_reaper.py's structural census of the
pending-fail predicate's one legitimate definition site is updated to
check sweep.py instead of router.py.
